[todoApp]jsを別ファイルに分離
